VIITE-2336 refactor
